Energy-monitor: android build and wasm simplicifation
This demo support both mcus, web, desktop and mobile.

Split in a lib.rs so that there are no warnings even if it means some
dummy wrapper in main.

Add instructions in the README
