+++ examples/energy-monitor/index.html
@@ -7,8 +7,7 @@
 <!--
   This is a static html file used to display the wasm build.
   In order to generate the build
-   - uncomment the #wasm# lines in Cargo.toml
-   - Run `wasm-pack build --release --target web` in this directory.
+   - Run `wasm-pack build --release --target web --no-default-features --features slint/default,chrono` in this directory.
 -->
 
 <head>
